refactor the controllers and services
