+++ service/src/main/java/org/cbioportal/service/GenericAssayEnrichmentService.java
@@ -2,17 +2,22 @@ package org.cbioportal.service;
 
 import org.cbioportal.model.EnrichmentType;
 import org.cbioportal.model.GenericAssayBinaryEnrichment;
+import org.cbioportal.model.GenericAssayCategoricalEnrichment;
 import org.cbioportal.model.MolecularProfileCaseIdentifier;
 import org.cbioportal.service.exception.MolecularProfileNotFoundException;
 
 import java.util.List;
 import java.util.Map;
 
-public interface GenericAssayBinaryDataService {
+public interface GenericAssayEnrichmentService {
     List<GenericAssayBinaryEnrichment> getGenericAssayBinaryEnrichments(
         String molecularProfileId,
         Map<String, List<MolecularProfileCaseIdentifier>> molecularProfileCaseSets,
         EnrichmentType enrichmentType)
         throws MolecularProfileNotFoundException;
 
+    List<GenericAssayCategoricalEnrichment> getGenericAssayCategoricalEnrichments(String molecularProfileId,
+                                                                                  Map<String, List<MolecularProfileCaseIdentifier>> molecularProfileCaseSets,
+                                                                                  EnrichmentType enrichmentType)
+        throws MolecularProfileNotFoundException;
 }
