+++ service/src/test/java/org/cbioportal/service/impl/GenericAssayEnrichmentServiceImplTest.java
@@ -1,0 +1,4 @@
+package org.cbioportal.service.impl;
+
+public class GenericAssayEnrichmentServiceImplTest {
+}
